pyproject.toml
[project]
name = "nig_law_ai"
version = "0.1.0"
description = "Nigerian Legal AI Assistant"
authors = [{ name = "Chukwuemeka" }]
requires-python = ">=3.10"

dependencies = [
  "streamlit==1.35.0",
  "openai==0.27.6",
  "langchain-core>=0.2.0,<0.3.0",
  "langchain-community>=0.0.24",
  "PyMuPDF==1.24.2",
  "pdfminer.six==20221105",
  "docx2txt==0.8",
  "python-docx==1.1.0",
  "python-dotenv==1.0.1",
  "pyttsx3==2.90",
  "pillow==9.5.0",
  "chardet==5.2.0",
  "pytesseract==0.3.10",
  "opencv-python==4.9.0.80",
  "httpx==0.27.0",
  "clerk-sdk-python==0.1.0",
  "psycopg2-binary==2.9.9",
  "SQLAlchemy==2.0.27",
  "faiss-cpu==1.7.4",
  "requests==2.31.0",
  "python-jose==3.3.0",
  "cryptography==42.0.2",
  "firebase-admin==6.4.0",
  "pyrebase4==4.8.0",
  "sentence-transformers",
  "pgvector",
  "PyPDF2==3.0.1",
  "numpy==1.26.4",
  "streamlit-chat",
  "faster-whisper",
  "streamlit-webrtc",
  "av",                  # required by streamlit-webrtc
  "scipy",               # for audio saving
]

[tool.uv]
# Optional settings can go here
